requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.20
youtube-transcript-api>=0.6.0
yt-dlp>=2024.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
httpx>=0.27.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
